+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.4 Release
+
+Ground up rewrite, uses the new generic class finder features from opal. Uses local
+storage to cache episode ids once the round has begun
+
 ### 0.2 Release
 
 Breaks backwards compatibility with Django < 1.7 (Test runner calls django.setup()